--- 03_URLSessionAndUICollection/Constants.swift
@@ -1,12 +1,0 @@
-//
-//  Constants.swift
-//  03_URLSessionAndUICollection
-//
-//  Created by FTS on 02/10/2023.
-//
-
-import Foundation
-
-struct Constants {
-
-}
