+++ 03_URLSessionAndUICollection/Utilities/ApiHandler.swift
@@ -75,4 +75,5 @@ class ApiHandler {
                 }
             }
     }
+    
 }
