+++ 03_URLSessionAndUICollection/Utilities/ErrorsModel.swift
