+++ 03_URLSessionAndUICollection/Views/FollowerCollectionViewCell.swift
@@ -17,15 +17,10 @@ class FollowerCollectionViewCell: UICollectionViewCell {
     @IBOutlet weak var followerImg: UIImageView!
     @IBOutlet weak var followerName: UILabel!
     
-    @IBOutlet weak var imgHeight: NSLayoutConstraint!
-    @IBOutlet weak var imgWidth: NSLayoutConstraint!
-    
     static var id = "followerCell"
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        followerImg.heightAnchor.constraint(equalTo: contentView.heightAnchor, multiplier: 0.55).isActive = true
-        followerImg.widthAnchor.constraint(equalTo: contentView.widthAnchor, multiplier: 0.55).isActive = true
     }
     
     func configureCell(model: FollowerCellModel) {
